Review #28 - JwtUtil 싱글톤으로 변경, removePrefix 를 AuthInterceptor 로 이동
